added dark godcat
