+++ Localization/en-US_Mods.EBF.NPCs.hjson
@@ -270,3 +270,5 @@ CannonTurret.DisplayName: Cannon Turret
 HarpoonTurret.DisplayName: Harpoon Turret
 NukeStand.DisplayName: Nuke Stand
 Godcat.DisplayName: Godcat
+Godcat_Light.DisplayName: Godcat_ Light
+Godcat_Dark.DisplayName: Godcat_ Dark
