+++ Localization/en-US_Mods.EBF.hjson
@@ -195,6 +195,9 @@ Projectiles: {
 	Godcat_LightBlade.DisplayName: Godcat_ Light Blade
 	Godcat_LightReturnBall.DisplayName: Godcat_ Light Return Ball
 	Godcat_LightBall.DisplayName: Godcat_ Light Ball
+	Godcat_DarkBall.DisplayName: Godcat_ Dark Ball
+	Godcat_DarkBlade.DisplayName: Godcat_ Dark Blade
+	Godcat_DarkReturnBall.DisplayName: Godcat_ Dark Return Ball
 }
 
 Items: {
